+++ src/sqlfluff/rules/std.py
@@ -57,6 +57,77 @@ class L005(BaseRule):
         return dict(previous_whitespace=(c.context == 'whitespace'))
 
 
+class L006(BaseRule):
+    """Operators should be surrounded by a single space unless at the start/end of a line"""
+
+    # Initialise the memory with a dict
+    init_m = dict(cm3=None, cm2=None, cm1=None)
+
+    @staticmethod
+    def eval_func(c, m):
+        cm1 = m['cm1']
+        if cm1:
+            context_m1 = cm1.context
+            if context_m1 == 'operator':
+                # The last chunk was an operator
+                # This is where the operator should kick in
+
+                # First check preceeding whitespace
+                cm2 = m['cm2']
+                if cm2:
+                    if cm2.context == 'whitespace':
+                        # it is whitespace (good)
+                        # check length and indentation (more than 1 allowed if it's an indent)
+                        if len(cm2.chunk) != 1 and cm2.start_pos != 0:
+                            # Preceeding whitespace is wrong length
+                            return True
+                    else:
+                        # Preceeding content is not whitespace
+                        return True
+
+                # Second check following whitespace
+                if c.context == 'whitespace':
+                    # it is whitespace (good)
+                    # check length
+                    if len(c.chunk) != 1:
+                        # Following whitespace is wrong length
+                        return True
+                else:
+                    # Following content is not whitespace
+                    return True
+        return False
+
+    @staticmethod
+    def memory_func(c, m):
+        # Here we want to keep a buffer of the previous 3 chunks.
+        # Entering this function, we want to first rotate the old ones
+        # and then add the new ones
+        return dict(
+            cm3=m.get('cm2', None),
+            cm2=m.get('cm1', None),
+            cm1=c)
+
+
+class L007(L006):
+    """Operators should be at the start of lines rather than the end"""
+
+    # This class is very similar to L006, but only check if an operator is followed by a newline
+    @staticmethod
+    def eval_func(c, m):
+        cm1 = m['cm1']
+        if cm1:
+            context_m1 = cm1.context
+            if context_m1 == 'operator':
+                # The last chunk was an operator
+                # This is where the operator should kick in
+
+                # Second check following whitespace
+                if '\n' in c.chunk:
+                    # Following whitespace contains a newline
+                    return True
+        return False
+
+
 class StandardRuleSet(BaseRuleSet):
     """ A standard set of SQL rules """
     rules = [
@@ -70,5 +141,5 @@ class StandardRuleSet(BaseRuleSet):
             'L003', "Single indentation uses a number of spaces not a multiple of 4",
             lambda c, m: c.context == 'whitespace' and c.start_pos == 0 and c.chunk.count(' ') % 4 != 0),
         # Defined a seperate rules
-        L004, L005
+        L004, L005, L006, L007
     ]